Move inline impls to src and update build
Extract many static member definitions and non-inline implementations from headers into new/updated src/*.cpp files (e.g. rcpp_interface_base.cpp, rcpp_*.cpp, fims_json.cpp). This reduces header bloat and avoids multiple-definition/linker issues by centralizing definitions. Remove JSON parser implementation from inst/include/utilities/fims_json.hpp and provide fims_json.cpp implementation. Remove TMB checks/library call from .onLoad (R/zzz.R). Update build settings in src/Makevars (comment PKG_CPPFLAGS lines, add CXX17STD/USE_CXX17). Also minor include/whitespace fixes and add placeholder fims_log.cpp.
