+++ inst/include/interface/rcpp/rcpp_objects/rcpp_growth.hpp
@@ -67,12 +67,7 @@ class GrowthInterfaceBase : public FIMSRcppInterfaceBase {
    */
   virtual double evaluate(double age) = 0;
 };
-// static id of the GrowthInterfaceBase object
-uint32_t GrowthInterfaceBase::id_g = 1;
-// local id of the GrowthInterfaceBase object map relating the ID of the
-// GrowthInterfaceBase to the GrowthInterfaceBase objects
-std::map<uint32_t, std::shared_ptr<GrowthInterfaceBase>>
-    GrowthInterfaceBase::live_objects;
+
 
 /**
  * @brief Rcpp interface for EWAAGrowth to instantiate the object from R:
