+++ inst/include/interface/rcpp/rcpp_objects/rcpp_interface_base.hpp
@@ -99,10 +99,7 @@ class Parameter {
     id_m = Parameter::id_g++;
   }
 };
-/**
- * @brief The unique ID for the variable map that points to a fims::Vector.
- */
-uint32_t Parameter::id_g = 0;
+
 
 /**
  * @brief Sanitize a double value by replacing NaN or Inf with -999.0.
@@ -124,7 +121,7 @@ inline double sanitize_val(double x) {
  * @param p A parameter.
  * @return std::ostream&
  */
-std::ostream& operator<<(std::ostream& out, const Parameter& p) {
+inline std::ostream& operator<<(std::ostream& out, const Parameter& p) {
   out << "{\"id\": " << p.id_m
       << ",\n\"value\": " << sanitize_val(p.initial_value_m)
       << ",\n\"estimated_value\": " << sanitize_val(p.final_value_m);
@@ -347,7 +344,7 @@ class ParameterVector {
     }
   }
 };
-uint32_t ParameterVector::id_g = 0;
+
 
 /**
  * @brief Output for std::ostream& for a ParameterVector.
@@ -356,7 +353,7 @@ uint32_t ParameterVector::id_g = 0;
  * @param v A ParameterVector.
  * @return std::ostream&
  */
-std::ostream& operator<<(std::ostream& out, ParameterVector& v) {
+inline std::ostream& operator<<(std::ostream& out, ParameterVector& v) {
   out << "[";
   size_t size = v.size();
   for (size_t i = 0; i < size - 1; i++) {
@@ -568,7 +565,7 @@ class RealVector {
     }
   }
 };
-uint32_t RealVector::id_g = 0;
+
 
 /**
  *@brief Base class for all interface objects.
@@ -647,7 +644,6 @@ class FIMSRcppInterfaceBase {
     return ss.str();
   }
 };
-std::vector<std::shared_ptr<FIMSRcppInterfaceBase>>
-    FIMSRcppInterfaceBase::fims_interface_objects;
+
 
 #endif
