+++ inst/include/interface/rcpp/rcpp_objects/rcpp_recruitment.hpp
@@ -80,12 +80,6 @@ class RecruitmentInterfaceBase : public FIMSRcppInterfaceBase {
    */
   virtual double evaluate_process(size_t pos) = 0;
 };
-// static id of the RecruitmentInterfaceBase object
-uint32_t RecruitmentInterfaceBase::id_g = 1;
-// local id of the RecruitmentInterfaceBase object map relating the ID of the
-// RecruitmentInterfaceBase to the RecruitmentInterfaceBase objects
-std::map<uint32_t, std::shared_ptr<RecruitmentInterfaceBase>>
-    RecruitmentInterfaceBase::live_objects;
 
 /**
  * @brief Rcpp interface for Beverton--Holt to instantiate from R:
