+++ inst/include/interface/rcpp/rcpp_objects/rcpp_shared_primitive.hpp
@@ -665,7 +665,7 @@ class SharedInt {
  * @param rhs The `SharedInt` object on the right-hand side.
  * @return A new `SharedInt` object containing the sum.
  */
-SharedInt operator+(const int& lhs, const SharedInt& rhs) {
+inline SharedInt operator+(const int& lhs, const SharedInt& rhs) {
   return SharedInt(lhs + rhs.get());
 }
 
@@ -681,7 +681,7 @@ SharedInt operator+(const int& lhs, const SharedInt& rhs) {
  * @param rhs The `SharedInt` object on the right-hand side.
  * @return A new `SharedInt` object containing the result of the subtraction.
  */
-SharedInt operator-(const int& lhs, const SharedInt& rhs) {
+inline SharedInt operator-(const int& lhs, const SharedInt& rhs) {
   return SharedInt(lhs - rhs.get());
 }
 
@@ -696,7 +696,7 @@ SharedInt operator-(const int& lhs, const SharedInt& rhs) {
  * @param rhs The `SharedInt` object on the right-hand side.
  * @return A new `SharedInt` object containing the product.
  */
-SharedInt operator*(const int& lhs, const SharedInt& rhs) {
+inline SharedInt operator*(const int& lhs, const SharedInt& rhs) {
   return SharedInt(lhs * rhs.get());
 }
 
@@ -711,7 +711,7 @@ SharedInt operator*(const int& lhs, const SharedInt& rhs) {
  * @param rhs The `SharedInt` object on the right-hand side.
  * @return A new `SharedInt` object containing the quotient.
  */
-SharedInt operator/(const int& lhs, const SharedInt& rhs) {
+inline SharedInt operator/(const int& lhs, const SharedInt& rhs) {
   return SharedInt(lhs / rhs.get());
 }
 
@@ -726,7 +726,7 @@ SharedInt operator/(const int& lhs, const SharedInt& rhs) {
  * @return `true` if the integer is less than the `SharedInt`'s value,
  * `false` otherwise.
  */
-bool operator<(const int& lhs, const SharedInt& rhs) {
+inline bool operator<(const int& lhs, const SharedInt& rhs) {
   return (lhs < rhs.get());
 }
 
@@ -741,7 +741,7 @@ bool operator<(const int& lhs, const SharedInt& rhs) {
  * @return `true` if the integer is less than or equal to the `SharedInt`'s
  * value, `false` otherwise.
  */
-bool operator<=(const int& lhs, const SharedInt& rhs) {
+inline bool operator<=(const int& lhs, const SharedInt& rhs) {
   return (lhs <= rhs.get());
 }
 
@@ -756,7 +756,7 @@ bool operator<=(const int& lhs, const SharedInt& rhs) {
  * @return `true` if the integer is greater than the `SharedInt`'s value,
  * `false` otherwise.
  */
-bool operator>(const int& lhs, const SharedInt& rhs) {
+inline bool operator>(const int& lhs, const SharedInt& rhs) {
   return (lhs > rhs.get());
 }
 
@@ -771,7 +771,7 @@ bool operator>(const int& lhs, const SharedInt& rhs) {
  * @return `true` if the integer is greater than or equal to the
  * `SharedInt`'s value, `false` otherwise.
  */
-bool operator>=(const int& lhs, const SharedInt& rhs) {
+inline bool operator>=(const int& lhs, const SharedInt& rhs) {
   return (lhs >= rhs.get());
 }
 
@@ -1448,7 +1448,7 @@ class SharedReal {
  * @param rhs The SharedReal object on the right-hand side.
  * @return A new SharedReal object containing the sum.
  */
-SharedReal operator+(const double& lhs, const SharedReal& rhs) {
+inline SharedReal operator+(const double& lhs, const SharedReal& rhs) {
   return SharedReal(lhs + rhs.get());
 }
 
@@ -1465,7 +1465,7 @@ SharedReal operator+(const double& lhs, const SharedReal& rhs) {
  * @return A new SharedReal object containing the result of the
  * subtraction.
  */
-SharedReal operator-(const double& lhs, const SharedReal& rhs) {
+inline SharedReal operator-(const double& lhs, const SharedReal& rhs) {
   return SharedReal(lhs - rhs.get());
 }
 
@@ -1480,7 +1480,7 @@ SharedReal operator-(const double& lhs, const SharedReal& rhs) {
  * @param rhs The SharedReal object on the right-hand side.
  * @return A new SharedReal object containing the product.
  */
-SharedReal operator*(const double& lhs, const SharedReal& rhs) {
+inline SharedReal operator*(const double& lhs, const SharedReal& rhs) {
   return SharedReal(lhs * rhs.get());
 }
 
@@ -1495,7 +1495,7 @@ SharedReal operator*(const double& lhs, const SharedReal& rhs) {
  * @param rhs The SharedReal object on the right-hand side.
  * @return A new SharedReal object containing the quotient.
  */
-SharedReal operator/(const double& lhs, const SharedReal& rhs) {
+inline SharedReal operator/(const double& lhs, const SharedReal& rhs) {
   return SharedReal(lhs / rhs.get());
 }
 
@@ -2047,7 +2047,7 @@ class SharedBoolean {
  * @return `true` if the boolean is less than the SharedBoolean's
  * value, `false` otherwise.
  */
-bool operator<(const bool& lhs, const SharedBoolean& rhs) {
+inline bool operator<(const bool& lhs, const SharedBoolean& rhs) {
   return (lhs < rhs.get());
 }
 
@@ -2062,7 +2062,7 @@ bool operator<(const bool& lhs, const SharedBoolean& rhs) {
  * @return `true` if the boolean is less than or equal to the
  * SharedBoolean's value, `false` otherwise.
  */
-bool operator<=(const bool& lhs, const SharedBoolean& rhs) {
+inline bool operator<=(const bool& lhs, const SharedBoolean& rhs) {
   return (lhs <= rhs.get());
 }
 
@@ -2077,7 +2077,7 @@ bool operator<=(const bool& lhs, const SharedBoolean& rhs) {
  * @return `true` if the boolean is greater than the SharedBoolean's
  * value, `false` otherwise.
  */
-bool operator>(const bool& lhs, const SharedBoolean& rhs) {
+inline bool operator>(const bool& lhs, const SharedBoolean& rhs) {
   return (lhs > rhs.get());
 }
 
@@ -2092,7 +2092,7 @@ bool operator>(const bool& lhs, const SharedBoolean& rhs) {
  * @return `true` if the boolean is greater than or equal to the
  * SharedBoolean's value, `false` otherwise.
  */
-bool operator>=(const bool& lhs, const SharedBoolean& rhs) {
+inline bool operator>=(const bool& lhs, const SharedBoolean& rhs) {
   return (lhs >= rhs.get());
 }
