+++ src/FIMS.cpp
@@ -5,7 +5,6 @@
 // #include "../inst/include/interface/rcpp/rcpp_interface.hpp"
 #include "../inst/include/interface/interface.hpp"
 #include "../inst/include/interface/TMB/init.hpp"
-#include "../inst/include/interface/rcpp/fims_modules.hpp"
 #include "../inst/include/common/model.hpp"
 
 /// @cond
