+++ src/rcpp_fleet.cpp
@@ -1,1 +1,8 @@
-// #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_fleet.hpp"
+ #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_fleet.hpp"
+
+ // static id of the FleetInterfaceBase object
+uint32_t FleetInterfaceBase::id_g = 1;
+// local id of the FleetInterfaceBase object map relating the ID of the
+// FleetInterfaceBase to the FleetInterfaceBase objects
+std::map<uint32_t, std::shared_ptr<FleetInterfaceBase>>
+    FleetInterfaceBase::live_objects;
