+++ src/rcpp_growth.cpp
@@ -1,1 +1,8 @@
-// #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_growth.hpp"
+ #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_growth.hpp"
+
+ // static id of the GrowthInterfaceBase object
+uint32_t GrowthInterfaceBase::id_g = 1;
+// local id of the GrowthInterfaceBase object map relating the ID of the
+// GrowthInterfaceBase to the GrowthInterfaceBase objects
+std::map<uint32_t, std::shared_ptr<GrowthInterfaceBase>>
+    GrowthInterfaceBase::live_objects;
