+++ src/rcpp_models.cpp
@@ -1,1 +1,8 @@
-// #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_models.hpp"
+ #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_models.hpp"
+
+ // static id of the FleetInterfaceBase object
+uint32_t FisheryModelInterfaceBase::id_g = 1;
+
+// FleetInterfaceBase to the FleetInterfaceBase objects
+std::map<uint32_t, std::shared_ptr<FisheryModelInterfaceBase>>
+    FisheryModelInterfaceBase::live_objects;
