+++ src/rcpp_natural_mortality.cpp
@@ -1,1 +1,1 @@
-// #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_natural_mortality.hpp"
+ #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_natural_mortality.hpp"
