+++ src/rcpp_shared_primitive.cpp
@@ -1,1 +1,1 @@
-// #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_shared_primitive.hpp"
+ #include "../inst/include/interface/rcpp/rcpp_objects/rcpp_shared_primitive.hpp"
